Create a typings file per source file, instead of per package.
This will allow you to pull in typings only for the parts of a package
you are using.
