+++ .gitignore
@@ -1,3 +1,3 @@
-node_modules
-lib
-src/test/fixtures
+/node_modules
+/lib
+/src/test/fixtures
