+++ package.json
@@ -10,20 +10,23 @@
   },
   "dependencies": {
     "@types/doctrine": "0.0.3",
+    "@types/fs-extra": "^4.0.2",
     "command-line-args": "^4.0.7",
     "command-line-usage": "^4.0.1",
     "doctrine": "^2.0.0",
     "escodegen": "^1.9.0",
+    "fs-extra": "^4.0.2",
     "polymer-analyzer": "^2.3.0"
   },
   "devDependencies": {
     "@types/chai": "^4.0.4",
+    "@types/glob": "^5.0.33",
     "@types/mocha": "^2.2.43",
     "bower": "^1.8.2",
     "chai": "^4.1.2",
     "clang-format": "^1.0.55",
+    "glob": "^7.1.2",
     "mocha": "^3.5.3",
-    "rimraf": "^2.6.2",
     "source-map-support": "^0.5.0",
     "typescript": "^2.5.3",
     "watchy": "^0.7.0"
