+++ src/test/goldens/paper-behaviors/paper-inky-focus-behavior.d.ts
@@ -1,0 +1,10 @@
+declare namespace Polymer {
+
+  /**
+   * `Polymer.PaperInkyFocusBehavior` implements a ripple when the element has keyboard focus.
+   */
+  interface PaperInkyFocusBehavior {
+    _createRipple(): any;
+    _focusedChanged(receivedFocusFromKeyboard: any): any;
+  }
+}
