+++ src/test/goldens/paper-behaviors/paper-ripple-behavior.d.ts
@@ -54,55 +54,4 @@ declare namespace Polymer {
     _createRipple(): PaperRippleElement;
     _noinkChanged(noink: any): any;
   }
-
-  interface PaperButtonBehavior {
-
-    /**
-     * The z-depth of this element, from 0-5. Setting to 0 will remove the
-     * shadow, and each increasing number greater than 0 will be "deeper"
-     * than the last.
-     */
-    elevation: number;
-    hostAttributes: Object|null;
-
-    /**
-     * In addition to `IronButtonState` behavior, when space key goes down,
-     * create a ripple down effect.
-     */
-    _spaceKeyDownHandler(event: KeyboardEvent): any;
-
-    /**
-     * In addition to `IronButtonState` behavior, when space key goes up,
-     * create a ripple up effect.
-     */
-    _spaceKeyUpHandler(event: KeyboardEvent): any;
-    _calculateElevation(): any;
-    _computeKeyboardClass(receivedFocusFromKeyboard: any): any;
-  }
-
-  /**
-   * `Polymer.PaperInkyFocusBehavior` implements a ripple when the element has keyboard focus.
-   */
-  interface PaperInkyFocusBehavior {
-    _createRipple(): any;
-    _focusedChanged(receivedFocusFromKeyboard: any): any;
-  }
-
-  /**
-   * Use `Polymer.PaperCheckedElementBehavior` to implement a custom element
-   * that has a `checked` property similar to `Polymer.IronCheckedElementBehavior`
-   * and is compatible with having a ripple effect.
-   */
-  interface PaperCheckedElementBehavior {
-
-    /**
-     * Synchronizes the element's `active` and `checked` state.
-     */
-    _buttonStateChanged(): any;
-
-    /**
-     * Synchronizes the element's checked state with its ripple effect.
-     */
-    _checkedChanged(): any;
-  }
 }
