+++ src/test/goldens/paper-button/paper-button.d.ts
