+++ src/test/goldens/polymer/lib/elements/custom-style.d.ts
@@ -1,0 +1,53 @@
+declare namespace Polymer {
+
+  /**
+   * Custom element for defining styles in the main document that can take
+   * advantage of [shady DOM](https://github.com/webcomponents/shadycss) shims
+   * for style encapsulation, custom properties, and custom mixins.
+   * 
+   * - Document styles defined in a `<custom-style>` are shimmed to ensure they
+   *   do not leak into local DOM when running on browsers without native
+   *   Shadow DOM.
+   * - Custom properties can be defined in a `<custom-style>`. Use the `html` selector
+   *   to define custom properties that apply to all custom elements.
+   * - Custom mixins can be defined in a `<custom-style>`, if you import the optional
+   *   [apply shim](https://github.com/webcomponents/shadycss#about-applyshim)
+   *   (`shadycss/apply-shim.html`).
+   * 
+   * To use:
+   * 
+   * - Import `custom-style.html`.
+   * - Place a `<custom-style>` element in the main document, wrapping an inline `<style>` tag that
+   *   contains the CSS rules you want to shim.
+   * 
+   * For example:
+   * 
+   * ```
+   * <!-- import apply shim--only required if using mixins -->
+   * <link rel="import href="bower_components/shadycss/apply-shim.html">
+   * <!-- import custom-style element -->
+   * <link rel="import" href="bower_components/polymer/lib/elements/custom-style.html">
+   * ...
+   * <custom-style>
+   *   <style>
+   *     html {
+   *       --custom-color: blue;
+   *       --custom-mixin: {
+   *         font-weight: bold;
+   *         color: red;
+   *       };
+   *     }
+   *   </style>
+   * </custom-style>
+   * ```
+   */
+  interface CustomStyle extends Polymer.Element {
+
+    /**
+     * Returns the light-DOM `<style>` child this element wraps.  Upon first
+     * call any style modules referenced via the `include` attribute will be
+     * concatenated to this element's `<style>`.
+     */
+    getStyle(): HTMLStyleElement|null;
+  }
+}
