+++ src/test/goldens/polymer/lib/elements/dom-module.d.ts
@@ -1,0 +1,32 @@
+declare namespace Polymer {
+
+  /**
+   * The `dom-module` element registers the dom it contains to the name given
+   * by the module's id attribute. It provides a unified database of dom
+   * accessible via its static `import` API.
+   * 
+   * A key use case of `dom-module` is for providing custom element `<template>`s
+   * via HTML imports that are parsed by the native HTML parser, that can be
+   * relocated during a bundling pass and still looked up by `id`.
+   * 
+   * Example:
+   * 
+   *     <dom-module id="foo">
+   *       <img src="stuff.png">
+   *     </dom-module>
+   * 
+   * Then in code in some other location that cannot access the dom-module above
+   * 
+   *     let img = customElements.get('dom-module').import('foo', 'img');
+   */
+  interface DomModule extends Polymer.Element {
+    attributeChangedCallback(name: any, old: any, value: any): any;
+
+    /**
+     * Registers the dom-module at a given id. This method should only be called
+     * when a dom-module is imperatively created. For
+     * example, `document.createElement('dom-module').register('foo')`.
+     */
+    register(id?: string): any;
+  }
+}
