+++ src/test/goldens/polymer/lib/utils/case-map.d.ts
@@ -1,0 +1,19 @@
+declare namespace Polymer {
+
+  namespace CaseMap {
+
+
+    /**
+     * Converts "dash-case" identifier (e.g. `foo-bar-baz`) to "camelCase"
+     * (e.g. `fooBarBaz`).
+     */
+    function dashToCamelCase(dash: string): string;
+
+
+    /**
+     * Converts "camelCase" identifier (e.g. `fooBarBaz`) to "dash-case"
+     * (e.g. `foo-bar-baz`).
+     */
+    function camelToDashCase(camel: string): string;
+  }
+}
