+++ src/test/goldens/polymer/lib/utils/flush.d.ts
@@ -1,0 +1,16 @@
+declare namespace Polymer {
+
+
+  /**
+   * Adds a `Polymer.Debouncer` to a list of globally flushable tasks.
+   */
+  function enqueueDebouncer(debouncer: Polymer.Debouncer|null): any;
+
+
+  /**
+   * Forces several classes of asynchronously queued tasks to flush:
+   * - Debouncers added via `enqueueDebouncer`
+   * - ShadyDOM distribution
+   */
+  function flush(): any;
+}
