+++ src/test/goldens/polymer/polymer-element.d.ts
@@ -1,0 +1,10 @@
+declare namespace Polymer {
+
+  /**
+   * Base class that provides the core API for Polymer's meta-programming
+   * features including template stamping, data-binding, attribute deserialization,
+   * and property change observation.
+   */
+  interface Element extends Polymer.Element {
+  }
+}
